% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_handler.R
\name{arg_spec}
\alias{arg_spec}
\title{Create an argument specification object}
\usage{
arg_spec(
  type = c("character", "integer", "logical", "numeric"),
  desc = "",
  optional = grepl("\\\\?$", type)
)
}
\arguments{
\item{type}{A string indicating the data type that is required for this
argument.}

\item{desc}{A human-readable description of the argument. Used to generate
documentation.}

\item{optional}{If \code{TRUE}, then this argument need not be present in a
request.}
}
\description{
Create an argument specification object
}
