% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_connect}
\alias{check_connect}
\title{Perform a series of checks to determine if the current code is executing on
RStudio Connect. This intentionally casts a wide net since there's no
persistent and reliable way to check of RStudio Connect is executing the R
session}
\usage{
check_connect()
}
\description{
Perform a series of checks to determine if the current code is executing on
RStudio Connect. This intentionally casts a wide net since there's no
persistent and reliable way to check of RStudio Connect is executing the R
session
}
