% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_route}
\alias{check_route}
\title{Check Plumber route for Tableau compliance}
\usage{
check_route(route)
}
\arguments{
\item{route}{A plumber route}
}
\value{
Provides warnings based on features of \code{route}
}
\description{
Check Plumber route for Tableau compliance
}
