% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_handler.R
\name{error_handler}
\alias{error_handler}
\title{Gracefully handle Plumber errors so they are effectively communicated back to Tableau}
\usage{
error_handler(req, res, err)
}
\arguments{
\item{req}{Plumber \code{req} object}

\item{res}{Plumber \code{res} object}

\item{err}{An error object passed from Plumber}
}
\description{
Gracefully handle Plumber errors so they are effectively communicated back to Tableau
}
