% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_tableau_request.R
\name{mock_tableau_request}
\alias{mock_tableau_request}
\title{Create a mock JSON request that mimics the request structure of Tableau}
\usage{
mock_tableau_request(script, data, pretty = TRUE, ...)
}
\arguments{
\item{script}{String indicating the endpoint path}

\item{data}{A list or dataframe that is serialized to JSON}

\item{pretty}{Whether or not to 'prettify' the JSON output}

\item{...}{Additional arguments passed to \code{jsonlite::toJSON()}}
}
\value{
A JSON object that can be passed to a Tableau endpoint
}
\description{
Create a mock JSON request that mimics the request structure of Tableau
}
\examples{
mock_tableau_request("/loess/predict", mtcars[,c("hp", "mpg")])

}
