% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbertableau.R
\docType{package}
\name{plumbertableau-package}
\alias{plumbertableau}
\alias{plumbertableau-package}
\title{plumbertableau: Extend Tableau with Plumber APIs}
\description{
Build and publish Plumber APIs that are compatible with the
    Tableau Analytics Extension API. Using annotations in R comments, extensions
    will respond to requests from Tableau, allowing you to call R code from
    Tableau workbooks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/plumbertableau/}
  \item \url{https://github.com/rstudio/plumbertableau}
  \item Report bugs at \url{https://github.com/rstudio/plumbertableau/issues}
}

}
\author{
\strong{Maintainer}: James Blair \email{james@rstudio.com}

Authors:
\itemize{
  \item Joe Cheng \email{joe@rstudio.com}
  \item Toph Allen \email{toph@rstudio.com}
  \item Bill Sager \email{bill.sager@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item Tableau [copyright holder]
}

}
\keyword{internal}
