% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_path}
\alias{standardize_path}
\title{Ensure path value starts with / and doesn't include double /}
\usage{
standardize_path(path)
}
\description{
Ensure path value starts with / and doesn't include double /
}
