% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_extension.R
\name{tableau_extension}
\alias{tableau_extension}
\title{Make an existing Plumber API compliant as a Tableau Analytics Extension}
\usage{
tableau_extension(path = "my-extension", warnings = TRUE)
}
\arguments{
\item{path}{The root path to use to mount existing Plumber endpoints to. This
allows local development to mirror the deployed API when it is deployed to
RStudio Connect.}

\item{warnings}{Whether or not to provide warnings about non-Tableau compliant
endpoints.}
}
\value{
A function that can be used with \code{#* @plumber}
}
\description{
Make an existing Plumber API compliant as a Tableau Analytics Extension
}
\examples{
\dontrun{
library(plumber)
library(plumbertableau)

#* Capitalize incoming text
#* @post /capitalize
function(req, res) {
  dat <- req$body$data
  toupper(dat)
}

#* @plumber
tableau_extension()
}

}
