% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_json.R
\name{tableau_json}
\alias{tableau_json}
\title{Create a JSON object that mimics a request from Tableau}
\usage{
tableau_json(script, data, ...)
}
\arguments{
\item{script}{String indicating the path to the endpoint to be called}

\item{data}{A list or dataframe that is serialized to JSON}

\item{...}{Additional arguments passed to \code{jsonlite::toJSON()}}
}
\value{
A JSON object that can be passed to a Tableau endpoint
}
\description{
\code{tableau_json()} creates a JSON object formatted like a request from Tableau.
The JSON object it returns can be pasted directly into the "Try it out" field
in the Swagger documentation for an endpoint to test its functionality.
}
\details{
Behind the scenes, Tableau sends all requests to the \verb{/evaluate} endpoint.
Each request is a JSON object containing two items: \code{script} and \code{data}.
plumbertableau uses \code{script} to specify an individual endpoint to call, and
passes the arguments in \code{data} on to the function at that endpoint.
}
