% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_request.R
\name{validate_request}
\alias{validate_request}
\title{Validate an incoming Tableau API request}
\usage{
validate_request(req, ...)
}
\arguments{
\item{req}{The \code{req} object of a Plumber request}

\item{...}{Named values to parse out of \code{req$body$data} and the expected
data type}
}
\value{
A list of named values parsed from \code{req$body$data}
}
\description{
Validate an incoming Tableau API request
}
\examples{
\dontrun{

}

}
