% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_route}
\alias{check_route}
\title{Checks a Plumber router for Tableau compliance: that all endpoints accept
POST requests and use the default JSON parser and serializer.}
\usage{
check_route(route)
}
\arguments{
\item{route}{A plumber route}
}
\value{
Provides warnings based on features of \code{route}
}
\description{
Checks a Plumber router for Tableau compliance: that all endpoints accept
POST requests and use the default JSON parser and serializer.
}
