% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_route}
\alias{check_route}
\title{Checks a Plumber route for Tableau compliance}
\usage{
check_route(route)
}
\arguments{
\item{route}{A plumber route}
}
\value{
Provides warnings based on features of \code{route}
}
\description{
Checks a route to ensure that it accepts POST requests and uses the default JSON parser and serializer.
}
\keyword{internal}
