% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbertableau-package.R
\docType{package}
\name{plumbertableau-package}
\alias{plumbertableau}
\alias{plumbertableau-package}
\title{plumbertableau: Create Tableau Analytics Extension API Compliant Plumber APIs}
\description{
The Tableau Analytics Extension API allows Tableau to be extended
    to external services. This package provides helpers for creating Plumber
    APIs that properly function as Tableau Analytics Extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/plumbertableau/}
  \item \url{https://github.com/rstudio/plumbertableau}
}

}
\author{
\strong{Maintainer}: James Blair \email{james@rstudio.com}

Other contributors:
\itemize{
  \item  RStudio [copyright holder, funder]
  \item  Tableau [copyright holder]
}

}
\keyword{internal}
