% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbertableau.R
\docType{package}
\name{plumbertableau-package}
\alias{plumbertableau}
\alias{plumbertableau-package}
\title{plumbertableau: Turn 'Plumber' APIs into 'Tableau' Extensions}
\description{
Build 'Plumber' APIs that can be used in 'Tableau' workbooks.
    Annotations in R comments allow APIs to conform to the 'Tableau Analytics
    Extension' specification, so that R code can be used to power 'Tableau'
    workbooks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/plumbertableau/}
  \item \url{https://github.com/rstudio/plumbertableau}
  \item Report bugs at \url{https://github.com/rstudio/plumbertableau/issues}
}

}
\author{
\strong{Maintainer}: James Blair \email{james@rstudio.com}

Authors:
\itemize{
  \item Joe Cheng \email{joe@rstudio.com}
  \item Toph Allen \email{toph@rstudio.com}
  \item Bill Sager \email{bill.sager@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item Tableau [copyright holder]
}

}
\keyword{internal}
