% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_extension.R
\name{tableau_extension}
\alias{tableau_extension}
\title{Make an existing Plumber API compliant as a Tableau Analytics Extension}
\usage{
tableau_extension(pr)
}
\arguments{
\item{pr}{A plumber router}
}
\value{
A modified plumber router that functions as a Tableau Analytics
Extension
}
\description{
Make an existing Plumber API compliant as a Tableau Analytics Extension
}
\examples{
\dontrun{
library(plumber)
library(plumbertableau)

#* Capitalize incoming text
#* @post /capitalize
function(req, res) {
  dat <- req$body$data
  toupper(dat)
}

#* @plumber
tableau_extension
}

}
