% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_extension.R
\name{tableau_extension}
\alias{tableau_extension}
\title{Make an existing Plumber API compliant as a Tableau Analytics Extension}
\usage{
tableau_extension(path = "/")
}
\arguments{
\item{path}{The root path to use to mount existing Plumber endpoints to}
}
\value{
A modified version of \code{pr} that is compliant with Tableau
}
\description{
Make an existing Plumber API compliant as a Tableau Analytics Extension
}
\examples{
\dontrun{
library(plumber)
library(rtab)

#* Capitalize incoming text
#* @parser json
#* @serializer json
#* @post /capitalize
function(req, res) {
  dat <- req$body$data
  toupper(dat)
}

#* @plumber
tableau_extension()
}

}
