% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_handler.R
\name{tableau_handler}
\alias{tableau_handler}
\title{Create a Tableau compliant handler}
\usage{
tableau_handler(args, func)
}
\arguments{
\item{args}{A named list detailing the expected names and types of incoming
data from Tableau}

\item{func}{A function to be used as the handler function}
}
\description{
Create a Tableau compliant handler
}
