% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_json.R
\name{tableau_json}
\alias{tableau_json}
\title{Create a mock JSON request that mimics the behavior of Tableau}
\usage{
tableau_json(script, data, ...)
}
\arguments{
\item{script}{String indicating the endpoint path}

\item{data}{A list or dataframe that is serialized to JSON}

\item{...}{Additional arguments passed to \code{jsonlite::toJSON()}}
}
\value{
A JSON object that can be passed to a Tableau endpoint
}
\description{
Create a mock JSON request that mimics the behavior of Tableau
}
